requirements.txt

streamlit>=1.30.0
python-dotenv>=1.0.0
mistralai>=1.6.0       
requests>=2.31.0
beautifulsoup4>=4.12.2
urllib3>=2.2.1